+++ src/format/mod.rs
@@ -26,6 +26,7 @@ pub(crate) mod world;
 
 pub(crate) use cursor::Cursor;
 pub use mesh::Mesh;
+pub(crate) use mesh::Pose;
 pub use world::World;
 
 use crate::error::{Error, Malformation, Result};